+++ CHANGELOG.md
@@ -45,6 +45,8 @@
 
 * `Set#superset?`, `Set#subset?`, and the respective 'proper_' variant of each are now implemented
 
+* Now with enabled arity checks calling a method with more arguments than those supported by its signature raises an `ArgumentError` as well.
+
 * `NameError` and `NoMethodError` - add `#name` and `#args` attributes
 
 ## 0.8.1 2015-10-12